Search Bar added int Pending & Completed + Search Bar script moved
