+++ dashboard/allStudents.php
@@ -7,7 +7,8 @@ if(!isset($_SESSION['userId'])){
 ?>
 <div id="content_table">
 	
-require('searchbar.php');
+<input type="text" id="myInput" onkeyup="myFunction()" style="" placeholder="Search for names..">
+    
 <p id="studentsList">List of Students Registered:</p>
 	<table id="myTable">
 		<tr>
@@ -41,25 +42,3 @@ require('searchbar.php');
 	</table>
 </div>
 
-<script>
-function myFunction() {
-  // Declare variables 
-  var input, filter, table, tr, td, i;
-  input = document.getElementById("myInput");
-  filter = input.value.toUpperCase();
-  table = document.getElementById("myTable");
-  tr = table.getElementsByTagName("tr");
-
-  // Loop through all table rows, and hide those who don't match the search query
-  for (i = 0; i < tr.length; i++) {
-    td = tr[i].getElementsByTagName("td")[0];
-    if (td) {
-      if (td.innerHTML.toUpperCase().indexOf(filter) > -1) {
-        tr[i].style.display = "";
-      } else {
-        tr[i].style.display = "none";
-      }
-    } 
-  }
-}
-</script>
