+++ dashboard/index.php
@@ -11,6 +11,8 @@ if(!isset($_SESSION['userId'])){
 	<link rel="stylesheet" href="css/adminTest.css">
 	<link rel="stylesheet" href="css/data.css">
 	<link rel="stylesheet" href="css/overview.css">
+    <link rel="stylesheet" href="css/searchbar.css">
+
 	
 	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
 	<script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
@@ -50,5 +52,26 @@ if(!isset($_SESSION['userId'])){
 		$('#content').load(loadData+".php");
 	}
 </script>
+<script>
+function myFunction() {
+  // Declare variables 
+  var input, filter, table, tr, td, i;
+  input = document.getElementById("myInput");
+  filter = input.value.toUpperCase();
+  table = document.getElementById("myTable");
+  tr = table.getElementsByTagName("tr");
 
+  // Loop through all table rows, and hide those who don't match the search query
+  for (i = 0; i < tr.length; i++) {
+    td = tr[i].getElementsByTagName("td")[0];
+    if (td) {
+      if (td.innerHTML.toUpperCase().indexOf(filter) > -1) {
+        tr[i].style.display = "";
+      } else {
+        tr[i].style.display = "none";
+      }
+    } 
+  }
+}
+</script>
 </html>
